Moved the PI list sort to the react-query query.
